+++ plugins/builtin/apm/nomad/plugin/job.go
@@ -68,9 +68,41 @@ func (a *APMPlugin) getTaskGroupResourceUsage(query *taskGroupQuery) ([]float64,
 		metricFunc = func(m *[]float64, ru *api.ResourceUsage) {
 			*m = append(*m, ru.CpuStats.Percent)
 		}
+	case queryMetricCPUAllocated:
+
+		// Since the Nomad API does not provide a metric for the percentage of CPU used
+		// out of amount allocated for taskgroups, the calculation must be done here.
+		// The total CPU allocated to the task group is retrieved once here since it
+		// does not vary between allocations.
+		allocatedCPU, err := a.getAllocatedCPUForTaskGroup(query.job, query.group)
+		if err != nil {
+			return nil, fmt.Errorf("failed to get total alloacted CPU for taskgroup: %v", err)
+		}
+
+		// Create the metric function now that the total allocated CPU is known
+		metricFunc = func(m *[]float64, ru *api.ResourceUsage) {
+			*m = append(*m, (ru.CpuStats.TotalTicks/float64(allocatedCPU))*100)
+		}
 	case queryMetricMem:
 		metricFunc = func(m *[]float64, ru *api.ResourceUsage) {
 			*m = append(*m, float64(ru.MemoryStats.Usage))
+		}
+	case queryMetricMemAllocated:
+
+		// Similarly to `queryMetricCPUAllocated` we must calculate the alloacted
+		// memory since it's not provided as a metric.
+		allocatedMem, err := a.getAllocatedMemForTaskGroup(query.job, query.group)
+		if err != nil {
+			return nil, fmt.Errorf("failed to get total alloacted memory for taskgroup: %v", err)
+		}
+
+		// Create the metric function now that the total allocated memory is known.
+		// The job info has total memory in memibytes (incorrectly labeled as MB) and
+		// the `MemoryStats.Usage` metric is reported in bytes, so we must convert one
+		// of them.
+		metricFunc = func(m *[]float64, ru *api.ResourceUsage) {
+			usageMiB := ru.MemoryStats.Usage / 1024 / 1024
+			*m = append(*m, (float64(usageMiB)/float64(allocatedMem))*100)
 		}
 	}
 
@@ -105,6 +137,55 @@ func (a *APMPlugin) getTaskGroupResourceUsage(query *taskGroupQuery) ([]float64,
 	}
 
 	return resp, nil
+}
+
+// getAllocatedCPUForTaskGroup calculates the total allocated CPU in MHz for a taskgroup
+func (a *APMPlugin) getAllocatedCPUForTaskGroup(job, taskgroup string) (int, error) {
+	taskGroupConfig, err := a.getTaskGroup(job, taskgroup)
+	if err != nil {
+		return -1, err
+	}
+
+	taskGroupAllocatedCPU := 0
+	for _, task := range taskGroupConfig.Tasks {
+		if task.Resources == nil || task.Resources.CPU == nil {
+			continue
+		}
+		taskGroupAllocatedCPU += *task.Resources.CPU
+	}
+	return taskGroupAllocatedCPU, nil
+}
+
+// getAllocatedMemForTaskGroup calculates the total allocated memory in MiB for a taskgroup
+func (a *APMPlugin) getAllocatedMemForTaskGroup(job, taskgroup string) (int, error) {
+	taskGroupConfig, err := a.getTaskGroup(job, taskgroup)
+	if err != nil {
+		return -1, err
+	}
+
+	taskGroupAllocatedMem := 0
+	for _, task := range taskGroupConfig.Tasks {
+		if task.Resources == nil || task.Resources.MemoryMB == nil {
+			continue
+		}
+		taskGroupAllocatedMem += *task.Resources.MemoryMB
+	}
+	return taskGroupAllocatedMem, nil
+}
+
+// getTaskGroup returns a task group configuration from a job.
+func (a *APMPlugin) getTaskGroup(job, taskgroup string) (*api.TaskGroup, error) {
+	jobInfo, _, err := a.client.Jobs().Info(job, nil)
+	if err != nil {
+		return nil, fmt.Errorf("failed to get info for job: %v", err)
+	}
+
+	taskGroupConfig := jobInfo.LookupTaskGroup(taskgroup)
+	if taskGroupConfig == nil {
+		return nil, fmt.Errorf("task group %q not found in job %q", taskgroup, job)
+	}
+
+	return taskGroupConfig, nil
 }
 
 // calculateTaskGroupResult determines the query result based on the metrics
@@ -169,7 +250,7 @@ func parseTaskGroupQuery(q string) (*taskGroupQuery, error) {
 	op := opMetricParts[1]
 	metric := opMetricParts[2]
 
-	if err := validateMetric(metric); err != nil {
+	if err := validateMetricTaskGroupQuery(metric); err != nil {
 		return nil, err
 	}
 	query.metric = metric
@@ -183,4 +264,8 @@ func parseTaskGroupQuery(q string) (*taskGroupQuery, error) {
 	}
 
 	return query, nil
+}
+
+func validateMetricTaskGroupQuery(metric string) error {
+	return validateMetric(metric, []string{queryMetricCPU, queryMetricCPUAllocated, queryMetricMem, queryMetricMemAllocated})
 }