+++ plugins/builtin/apm/nomad/plugin/job_test.go
@@ -66,10 +66,32 @@ func Test_parseTaskGroupQuery(t *testing.T) {
 			expectError: false,
 		},
 		{
+			name:  "avg_cpu-allocated",
+			input: "taskgroup_avg_cpu-allocated/group/job",
+			expected: &taskGroupQuery{
+				metric:    "cpu-allocated",
+				job:       "job",
+				group:     "group",
+				operation: "avg",
+			},
+			expectError: false,
+		},
+		{
 			name:  "avg_memory",
 			input: "taskgroup_avg_memory/group/job",
 			expected: &taskGroupQuery{
 				metric:    "memory",
+				job:       "job",
+				group:     "group",
+				operation: "avg",
+			},
+			expectError: false,
+		},
+		{
+			name:  "avg_memory-allocated",
+			input: "taskgroup_avg_memory-allocated/group/job",
+			expected: &taskGroupQuery{
+				metric:    "memory-allocated",
 				job:       "job",
 				group:     "group",
 				operation: "avg",