+++ plugins/builtin/apm/nomad/plugin/metrics_test.go
@@ -9,30 +9,46 @@ import (
 
 func Test_validateMetric(t *testing.T) {
 	testCases := []struct {
-		inputMetric    string
-		expectedOutput error
-		name           string
+		inputMetric       string
+		inputValidMetrics []string
+		expectedOutput    error
+		name              string
 	}{
 		{
-			inputMetric:    "memory",
-			expectedOutput: nil,
-			name:           "memory metric",
+			inputMetric:       "memory",
+			inputValidMetrics: []string{queryMetricCPU, queryMetricCPUAllocated, queryMetricMem, queryMetricMemAllocated},
+			expectedOutput:    nil,
+			name:              "memory metric",
 		},
 		{
-			inputMetric:    "cpu",
-			expectedOutput: nil,
-			name:           "cpu metric",
+			inputMetric:       "memory-allocated",
+			inputValidMetrics: []string{queryMetricCPU, queryMetricCPUAllocated, queryMetricMem, queryMetricMemAllocated},
+			expectedOutput:    nil,
+			name:              "memory-allocated metric",
 		},
 		{
-			inputMetric:    "cost-of-server",
-			expectedOutput: errors.New("invalid metric \"cost-of-server\", allowed values are cpu or memory"),
-			name:           "invalid metric",
+			inputMetric:       "cpu",
+			inputValidMetrics: []string{queryMetricCPU, queryMetricCPUAllocated, queryMetricMem, queryMetricMemAllocated},
+			expectedOutput:    nil,
+			name:              "cpu metric",
+		},
+		{
+			inputMetric:       "cpu-allocated",
+			inputValidMetrics: []string{queryMetricCPU, queryMetricCPUAllocated, queryMetricMem, queryMetricMemAllocated},
+			expectedOutput:    nil,
+			name:              "cpu-allocated metric",
+		},
+		{
+			inputMetric:       "cost-of-server",
+			inputValidMetrics: []string{queryMetricCPU, queryMetricCPUAllocated, queryMetricMem, queryMetricMemAllocated},
+			expectedOutput:    errors.New("invalid metric \"cost-of-server\", allowed values are: cpu, cpu-allocated, memory, memory-allocated"),
+			name:              "invalid metric",
 		},
 	}
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			actualOutput := validateMetric(tc.inputMetric)
+			actualOutput := validateMetric(tc.inputMetric, tc.inputValidMetrics)
 			assert.Equal(t, tc.expectedOutput, actualOutput, tc.name)
 		})
 	}